fix: fix check input to update the complete state of the task and refactor the code to the app.tsx
