+++ src/components/TaskList.tsx
@@ -1,46 +1,30 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { useForm } from "react-hook-form";
 import "./TaskList.css";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { searchTodoSchema } from "../validation/dataSchema";
 import { SearchFormData, Task } from "../types";
 
-const TaskList = () => {
+const TaskList = ({
+  tasks,
+  deleteTask,
+  handleCheckboxChange,
+}: {
+  tasks: Task[];
+  deleteTask: (task: Task) => void;
+  handleCheckboxChange: (id: string | undefined) => void;
+}) => {
   const { register } = useForm<SearchFormData>({
     resolver: zodResolver(searchTodoSchema),
   });
-
-  const [tasks, setTasks] = useState<Task[]>([]);
   const [searchValue, setSearchValue] = useState<string>("");
-
-  useEffect(() => {
-    const existingTasks = JSON.parse(localStorage.getItem("tasks") || "[]");
-    const tasksWithIds = existingTasks.map((task: Task) =>
-      task.id ? task : { ...task, id: `${Date.now()}-${Math.random()}` }
-    );
-    setTasks(tasksWithIds);
-  }, []);
 
   const filteredTasks = tasks.filter((task) =>
     task.taskName.toLowerCase().includes(searchValue)
   );
 
-  const handleDeleteTask = (id: string | undefined) => {
-    const updatedTasks = tasks.filter((task) => task.id !== id);
-    setTasks(updatedTasks);
-    localStorage.setItem("tasks", JSON.stringify(updatedTasks));
-  };
-
   const handleSearchChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setSearchValue(e.target.value.toLowerCase());
-  };
-
-  const handleCheckboxChange = (id: string | undefined) => {
-    const updatedTasks = tasks.map((task) =>
-      task.id === id ? { ...task, completed: !task.completed } : task
-    );
-    setTasks(updatedTasks);
-    localStorage.setItem("tasks", JSON.stringify(updatedTasks));
   };
 
   return (
@@ -73,7 +57,7 @@ const TaskList = () => {
                 />
                 <p className="taskCard-title">{task.taskName}</p>
               </div>
-              <button onClick={() => handleDeleteTask(task.id)}>Delete</button>
+              <button onClick={() => deleteTask(task)}>Delete</button>
             </li>
           ))}
         </ul>
